Now am able to insert gist like html into the editor
