+++ package.json
@@ -28,7 +28,6 @@
     "core-js": "^2.5.1",
     "hammerjs": "^2.0.8",
     "ng2-smart-table": "^1.2.2",
-    "node-gist-html": "^1.0.9",
     "rxjs": "^5.4.3",
     "zone.js": "^0.8.17"
   },
