+++ src/app/_models/Gist.ts
@@ -1,0 +1,5 @@
+export class Gist {
+  html: string;
+  file: string;
+  stylesheet: string;
+}
